+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/AndroidComponentsConfig.kt
@@ -209,6 +209,7 @@ fun ApplicationAndroidComponentsExtension.configure(
             project,
             "${variant.name}RuntimeClasspath",
             variant.name.capitalized,
+            variant.artifacts.get(SingleArtifact.MERGED_MANIFEST),
           )
         val javaSources = variant.sources.java
         if (buildTimeOptionsTask != null && javaSources != null) {